Increased screen border to make the ship more visible
Comments too.

+++ AmazonCodingExcercise/asteroids/code/asteroids.h
@@ -20,8 +20,10 @@ public:
     void Draw();
     void DrawImages();
     void DrawText();
+
     void OnStart();
     void OnUpdate(const float deltaTime);
+    
     void UserInput(const float deltaTime);
     void Fire(const float deltaTime);
     void UpdateThrust(const float deltaTime);
@@ -43,11 +45,13 @@ private:
     std::list<Object> enemies;
     std::list<TimedImage> timedImages;
 
+    // Text labels.
     const char scoreString[10] = "Score: %d";
     const char livesString[10] = "Lives: %d";
     const char thrustString[21] = "Thrust: %d%% (%.1fs)";
     const char explosionImageString[19] = "explosion_00%d.png";
 
+    // Player ship firing.
     const float fireDelay = .5f;
     float fireDelayTimer = .0f;
     const float fastFireDelay = .1f;
@@ -55,20 +59,23 @@ private:
     const float fastFireChargeDelay = 5.0f;
     float fastFireChargeDelayTimer = .0f;
 
-    const short maxLives = 3;
-    short lives = maxLives;
-    short score = 0;
+    // Player stats.
+    const uint8_t maxLives = 3;
+    int8_t lives = maxLives;
+    uint16_t score = 0;
 
-    const short normalMoveVelocity = 1;
-    const short thrustMoveVelocity = 3;
-    short moveVelocity = normalMoveVelocity;
+    // Player ship movement.
+    const uint8_t normalMoveVelocity = 1;
+    const uint8_t thrustMoveVelocity = 3;
+    uint8_t moveVelocity = normalMoveVelocity;
     const float thrustTime = 1.f;
     float thrustTimeTimer = .0f;
     const float thrustRechargeDelay = 5.f;
     float thrustRechargeDelayTimer = .0f;
     float prevThrustTime = 1000.f;
 
-    Uint8 hoverDirection = 0; // Left, Top, Right, Bottom.
+    // Player ship hover effect.
+    uint8_t hoverDirection = 0; // Left, Top, Right, Bottom.
     const float hoverDirectionDelay = .25f;
     float hoverDirectionDelayTimer = .0f;
 };
